Implement adding of component listener in world configuration
The component listener will be added together with registering the
component in the world configuration. That was done because when
registering the component listener the base type of the component is
needed. Without reflections that is not possible to detect it from the
component listener template type. Thus, adding component and listener
objects together is the most elegant workaround - I guess.